requirements.txt
beautifulsoup4==4.10.0
loguru==0.6.0
lxml==4.8.0
python-docx==0.8.11
requests==2.27.1
urllib3==1.26.8